+++ src/pages/Bookmark/BookmarkPage.jsx
@@ -1,0 +1,17 @@
+import React from "react";
+
+import Layout from "../../components/layout/Layout";
+
+import BookmarkLayout from "./_base/BookmarkLayout";
+
+const BookmarkPage = () => {
+  return (
+    <>
+      <Layout>
+        <BookmarkLayout></BookmarkLayout>
+      </Layout>
+    </>
+  );
+};
+
+export default BookmarkPage;
