--- src/pages/BookmarkPage.jsx
@@ -1,17 +1,0 @@
-import React from "react";
-
-import Layout from "../components/Layout/Layout";
-
-import BookmarkLayout from "../components/Bookmark/BookmarkLayout";
-
-const BookmarkPage = () => {
-  return (
-    <>
-      <Layout>
-        <BookmarkLayout></BookmarkLayout>
-      </Layout>
-    </>
-  );
-};
-
-export default BookmarkPage;
