+++ src/pages/Detail/DetailPage.jsx
@@ -1,0 +1,16 @@
+import React from "react";
+
+import Layout from "../../components/layout/Layout";
+import DetailLayout from "./_base/DetailLayout";
+
+const DetailPage = () => {
+  return (
+    <>
+      <Layout>
+        <DetailLayout></DetailLayout>
+      </Layout>
+    </>
+  );
+};
+
+export default DetailPage;
