--- src/pages/DetailPage.jsx
@@ -1,16 +1,0 @@
-import React from "react";
-
-import Layout from "../components/Layout/Layout";
-import DetailLayout from "../components/Detail/DetailLayout";
-
-const DetailPage = () => {
-  return (
-    <>
-      <Layout>
-        <DetailLayout></DetailLayout>
-      </Layout>
-    </>
-  );
-};
-
-export default DetailPage;
