+++ src/pages/Main/MainPage.jsx
@@ -1,0 +1,16 @@
+import React from "react";
+
+import Layout from "../../components/layout/Layout";
+import PostLayout from "./_base/PostLayout";
+
+const MainPage = () => {
+  return (
+    <>
+      <Layout>
+        <PostLayout></PostLayout>
+      </Layout>
+    </>
+  );
+};
+
+export default MainPage;
