--- src/pages/MainPage.jsx
@@ -1,16 +1,0 @@
-import React from "react";
-
-import Layout from "../components/Layout/Layout";
-import PostLayout from "../components/Post/PostLayout";
-
-const MainPage = () => {
-  return (
-    <>
-      <Layout>
-        <PostLayout></PostLayout>
-      </Layout>
-    </>
-  );
-};
-
-export default MainPage;
