+++ src/pages/Profile/ProfilePage.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+import Layout from "../../components/layout/Layout";
+import ProfileLayout from "./_base/ProfileLayout";
+
+const ProfilePage = () => {
+  return (
+    <Layout>
+      <ProfileLayout />
+    </Layout>
+  );
+};
+
+export default ProfilePage;
