--- src/pages/ProfilePage.jsx
@@ -1,14 +1,0 @@
-import React from "react";
-
-import Layout from "../components/Layout/Layout";
-import ProfileLayout from "../components/Profile/ProfileLayout";
-
-const ProfilePage = () => {
-  return (
-    <Layout>
-      <ProfileLayout />
-    </Layout>
-  );
-};
-
-export default ProfilePage;
